@@ -1,11 +1,12 @@
-import {Response} from '@angular/http'
+import {HttpErrorResponse} from '@angular/common/http'
 import {Observable} from 'rxjs/Observable'
 export class ErrorHandler{
   static handleError(error: Response | any){
     let errorMessage: string
 
-    if(error instanceof Response){
-      error = `Erro ${error.status} ao obter a url ${error.url} - ${error.statusText} `
+    if(error instanceof HttpErrorResponse){
+      const body = error.error
+      errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''} ${body}`
     }else{
       errorMessage = error.toString()
     }
