@@ -2,14 +2,14 @@ import {Injectable} from '@angular/core'
 import {ShoppingCartService} from '../restaurant-detail/shopping-cart/shopping-cart.service'
 import {CartItem} from '../restaurant-detail/shopping-cart/cart-item.model'
 import {Order,OrderItem} from './order.model'
-import {Http, Headers, RequestOptions} from '@angular/http'
+import {HttpClient} from '@angular/common/http'
 import {Observable} from 'rxjs/Observable'
 import 'rxjs/add/operator/map'
 
 import {MEAT_API} from '../app.api'
 @Injectable()
 export class OrderService{
-  constructor(private cartService:ShoppingCartService, private http: Http){
+  constructor(private cartService:ShoppingCartService, private http: HttpClient){
 
   }
   carItems(): CartItem[]{
@@ -33,14 +33,8 @@ export class OrderService{
 
   checkOrder(order: Order): Observable<string>{
     console.log(order)
-    const headers = new Headers()
-    headers.append('Content-Type','aplication/json')
-    const url = `${MEAT_API}/orders/${JSON.stringify(order)}`
-    console.log(`a url${url}`)
-    return this.http.post(`${MEAT_API}/orders`,
-      order,
-      new RequestOptions({headers: headers}))
-      .map(response=> response.json())
-      .map(order => order.id)
+ 
+    return this.http.post<Order>(`${MEAT_API}/orders`,
+      order).map(order => order.id)
     }
   }
