@@ -2,6 +2,17 @@ import { Component, OnInit } from '@angular/core';
 import {Restaurant} from './restaurant/restaurant.model'
 import {RestaurantsService} from './restaurants.service'
 import {trigger,state,style,transition,animate} from '@angular/animations'
+import {FormBuilder,FormControl, FormGroup} from '@angular/forms'
+import 'rxjs/add/operator/switchMap'
+import 'rxjs/add/operator/debounceTime'
+import 'rxjs/add/operator/distinctUntilChanged'
+import 'rxjs/add/operator/catch'
+import 'rxjs/add/observable/from'
+
+import {Observable}  from 'rxjs/Observable'
+
+
+
 
 @Component({
   selector: 'mt-restaurants',
@@ -25,13 +36,28 @@ import {trigger,state,style,transition,animate} from '@angular/animations'
 })
 export class RestaurantsComponent implements OnInit {
 
-  searchBarState = 'hidden'
+searchBarState = 'hidden'
 restaurants : Restaurant[]
+searchForm:FormGroup
+searchControl: FormControl
 
-  constructor(private restaurantService: RestaurantsService) { }
+  constructor(private restaurantService: RestaurantsService,
+  private fb: FormBuilder) { }
 
   ngOnInit() {
-     this.restaurantService.restaurants()
+    this.searchControl = this.fb.control('')
+    this.searchForm = this.fb.group({
+      searchControl: this.searchControl
+    })
+    this.searchControl.valueChanges
+    .debounceTime(500)
+    .distinctUntilChanged()
+    .switchMap(searchTerm =>  
+      this.restaurantService.restaurants(searchTerm)
+       .catch(error=>Observable.from([])))
+      .subscribe(restaurants => this.restaurants = restaurants)
+     
+      this.restaurantService.restaurants()
     .subscribe(restaurants => this.restaurants = restaurants)
   }
   toggleSearch(){
