@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core'
-import {Http} from '@angular/http'
+import {HttpClient,HttpParams} from '@angular/common/http'
 import {Restaurant} from "./restaurant/restaurant.model"
 import {MEAT_API} from '../app.api'
 import {Observable} from 'rxjs/Observable'
@@ -7,30 +7,32 @@ import 'rxjs/add/operator/map'
 import 'rxjs/add/operator/catch'
 import {ErrorHandler} from '../app.error-handler'
 import {MenuItem} from "../restaurant-detail/menu-item/menu-item.model"
+
 @Injectable()
 export class RestaurantsService{
 
-  constructor(private http:Http){
+  constructor(private http:HttpClient){
 
   }
-  restaurants(): Observable<Restaurant[]>{
-  return this.http.get(`${MEAT_API}/restaurants`)
-  .map(response => response.json())
-  .catch(ErrorHandler.handleError)
+  restaurants(search?: string): Observable<Restaurant[]>{
+    let params: HttpParams = undefined
+    if(search){
+      params = new HttpParams().append('q',search)
+    }
+  return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`,{params: params})
+ 
   }
   restaurantById(id: string): Observable<Restaurant>{
-    return this.http.get(`${MEAT_API}/restaurants/${id}`)
-    .map(response => response.json())
-    .catch(ErrorHandler.handleError)
+    return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`)
+   
   }
   reviewsOfRestaurant(id: string): Observable<any>{
     return this.http.get(`${MEAT_API}/restaurant/${id}/reviews`)
-    .map(response => response.json())
-    .catch(ErrorHandler.handleError)
+    
   }
   menuOfRestaurants(id: string): Observable<MenuItem[]>{
-    return this.http.get(`${MEAT_API}/restaurant/${id}/menu`).
-    map(response => response.json())
-    .catch(ErrorHandler.handleError)
+    return this.http.get<MenuItem[]>(`${MEAT_API}/restaurant/${id}/menu`)
+   
+    
   }
 }
